requirements.txt
aniso8601==9.0.1
boost-histogram==1.0.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
scipy==1.6.3
six==1.16.0
urllib3==1.26.5
urlopen==1.0.0
Werkzeug==2.0.1
